add mui, 필요없는 파일들 정리, Header 스타일 분리
