--- frontend/src/components/test/Task.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export default function Task({ task: { id, title, state }, onArchiveTask, onPinTask }) {
-  return (
-    <div className="list-item">
-      <input type="text" value={title} readOnly={true} />
-    </div>
-  );
-}
